chore: replace lib/time with lib/time-dayjs where it is possible
