+++ frontend/src/metabase-lib/v1/parameters/utils/click-behavior.ts
@@ -5,7 +5,7 @@ import {
   formatDateToRangeForParameter,
 } from "metabase/lib/formatting/date";
 import type { ValueAndColumnForColumnNameDate } from "metabase/lib/formatting/link";
-import { parseTimestamp } from "metabase/lib/time";
+import { parseTimestamp } from "metabase/lib/time-dayjs";
 import { checkNotNull } from "metabase/lib/types";
 import type { ClickObjectDimension as DimensionType } from "metabase-lib";
 import * as Lib from "metabase-lib";
