+++ frontend/src/metabase/models/components/ModelActions/ModelActionDetails/ModelActionDetails.tsx
@@ -8,7 +8,7 @@ import Actions from "metabase/entities/actions";
 import Databases from "metabase/entities/databases";
 import { useConfirmation } from "metabase/hooks/use-confirmation";
 import { connect } from "metabase/lib/redux";
-import { parseTimestamp } from "metabase/lib/time";
+import { parseTimestamp } from "metabase/lib/time-dayjs";
 import * as Urls from "metabase/lib/urls";
 import type Question from "metabase-lib/v1/Question";
 import {
