+++ frontend/src/metabase/notifications/PulsesListSidebar.jsx
@@ -17,7 +17,7 @@ import {
 } from "metabase/lib/formatting";
 import { getActivePulseParameters } from "metabase/lib/pulse";
 import { connect } from "metabase/lib/redux";
-import { formatFrame } from "metabase/lib/time";
+import { formatFrame } from "metabase/lib/time-dayjs";
 import { formatDateValue } from "metabase/parameters/utils/date-formatting";
 import { Button, Icon, Tooltip } from "metabase/ui";
